metal : add ggml_metal_library_t
ggml-ci
