+++ ggml/src/ggml-metal/CMakeLists.txt
@@ -10,7 +10,6 @@ ggml_add_backend_library(ggml-metal
                          ggml-metal-device.cpp
                          ggml-metal-common.cpp
                          ggml-metal-context.m
-                         ggml-metal-context.cpp
                          ggml-metal-ops.cpp
                         )
